requirements.txt
Flask==2.3.2
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
Werkzeug==2.3.6
SQLAlchemy==2.0.15
psycopg2-binary==2.9.6
python-dotenv==1.0.0
email-validator==2.0.0
WTForms==3.0.1
Flask-Migrate==4.0.4
Flask-Bcrypt==1.0.1
python-dateutil==2.8.2
PyJWT==2.6.0
requests==2.31.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-api-python-client==2.93.0
openpyxl==3.1.5
XlsxWriter==3.2.3 